# swagger.yml
openapi: 3.0.0
info:
  description: This is an API for the project on GitHub https://github.com/PandelisT/my-mood-diary.
  version: "1.0.0"
  title: My Mood Diary API
  contact:
    email: pandeli@nerdypandy.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/profile/{user_id}/image':
    get:
      tags:
        - profile
      summary: Logs user into the system
      operationId: getProfileImage
      parameters:
        - name: user_id
          in: path
          description: The user_id connected to the image
          required: true
          schema:
            type: string
        - name: firstname
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
    post:
      tags:
        - profile
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createProfileImage
      responses:
        default:
          description: successful operation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
        description: Created user object
        required: true
    put:
      tags:
        - profile
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateProfileImage
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
        description: Updated user object
        required: true
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        filename:
          type: string
      xml:
        name: Client
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/nerdypan/my-mood-diary/1.0.0